+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v1.0.3
+
+* GoogleApi::Ga::Data has now silent error, GoogleApi::Ga::Data.error(true) - raise error, GoogleApi::Ga::Data.error(false) - not error (default)
+* add GoogleApi::Ga::Data.use_cache
+* add GoogleApi::Ga::Data.clear_cache
+
 ## v1.0.2
 
 * ids is convert to integer